Move rendering types to render::types
